refactor: consolidate api routes under app directory
